refactor(desktop): extract markdown message link pill
Move the message link pill renderer into the markdown submodule without changing its implementation.
